nh-search: extract offline search flow
