+++ crates/nh-search/src/lib.rs
@@ -2,6 +2,7 @@ pub mod args;
 
 mod backend;
 mod channel;
+mod offline;
 mod online;
 mod query;
 mod render;
